Right align text for numbers. Kill worker thread.
Rather than have text be off, add statictext labels, that are right justified so the colons line up.

Numbers are now right aligned too, so that they line up and don't move.

Fix a problem with exit, where we need to kill the worker thread, and kill the timer.  This is done on cancel.